[red-knot] Move relation methods from `CallableType` to `Signature`
